requirement.txt
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
gmpy2==2.1.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
requests==2.32.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
triton==3.0.0
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
zipp==3.19.2